missionaries_cannibals v0.0.1 MC

Functions

create_worker(path)

Takes a path of states creates a new task to continue searching for solution.

start(type, args)

Callback implementation for c::application.start/2.

start_state()

Starts search by creating a task with the default start state.