missionaries_cannibals v0.0.1 MC.TaskWorker

Functions

duplicate?(state)

Checks if a state has already been used. Adds it to the public ets if it is new Returns boolean declaring if it has been perviously used.

goal?()

Tests if state is goal

start(list)

Takes the full path as a parameter. Uses helper functions to continue the search to the goal.

state_key(arg)

Takes a state tuple and converts it into a unique string.

valid?()

Takes a state and returns a boolean declaring if the state is valid.